+++ src/back/Plants.Application/Requests/SearchRequest.cs
@@ -6,12 +6,12 @@ public record SearchRequest(string? PlantName,
     decimal? LowerPrice,
     decimal? TopPrice,
     DateTime? LastDate,
-    int[]? GroupIds,
-    int[]? RegionIds,
-    int[]? SoilIds) : IRequest<SearchResult>;
+    long[]? GroupIds,
+    long[]? RegionIds,
+    long[]? SoilIds) : IRequest<SearchResult>;
 
 public record SearchResult(List<SearchResultItem> Items);
-public record SearchResultItem(int Id, string PlantName, string Description, int[] ImageIds, double Price)
+public record SearchResultItem(long Id, string PlantName, string Description, long[] ImageIds, double Price)
 {
     //used by converter
     public SearchResultItem() : this(0, "", "", null, 0)
